Move binary version tests to tools due to cross-build problems in sbtScalaNative
