+++ build.sbt
@@ -312,7 +312,7 @@ lazy val tools =
     .settings(
       libraryDependencies ++= Seq(
         "org.scalacheck" %% "scalacheck" % "1.14.3" % "test",
-        "org.scalatest" %% "scalatest" % "3.1.1" % "test"
+        "org.scalatest"  %% "scalatest"  % "3.1.1"  % "test"
       ),
       Test / fork := true,
       Test / javaOptions ++= {
